Added missing typecast.

+++ src/Magick.Native/Optimizers/JpegOptimizer.c
@@ -344,7 +344,7 @@ static size_t DetermineQuality(j_decompress_ptr decompress_info)
     }
   }
 
-  return quality < 85 ? quality : 85;
+  return (size_t) quality < 85 ? quality : 85;
 }
 
 static boolean DecompressJpeg(j_decompress_ptr decompress_info, ClientData *client_data)
